Fix type limits / absolute value warnings

+++ prboom2/src/gl_clipper.c
@@ -345,7 +345,7 @@ angle_t gld_FrustumAngle(void)
   }
 
   // If the pitch is larger than this you can look all around at a FOV of 90
-  if (D_abs(viewpitch) > 46 * ANG1)
+  if (D_abs((int) viewpitch) > 46 * ANG1)
     return 0xffffffff;
 
   // ok, this is a gross hack that barely works...
